chore(react): separate module federation tests into separate test files for better distribution
