+++ e2e/react/src/module-federation/utils.ts
@@ -1,0 +1,12 @@
+import { readJson } from '@nx/e2e/utils';
+import { join } from 'path';
+
+export function readPort(appName: string): number {
+  let config;
+  try {
+    config = readJson(join('apps', appName, 'project.json'));
+  } catch {
+    config = readJson(join(appName, 'project.json'));
+  }
+  return config.targets.serve.options.port;
+}
